Store posts in Database.
Posts that already have an ID entry will be skipped. Still need to write the code for making the post, however.

Added a --only-database option that is meant to skip posts and only store the
records. So you can just roll up a database
and then start posting new entries from
the feed.